Enhance account creation and validation
This commit enhances the process of creating new accounts and the associated validation. In particular, it adds "not blank" validation to the 'userName' and 'email' fields in 'AccountDTO'. Also, in 'AccountService', it sets id to null and status to active for every new account created. In addition, it alters account sequence in the import.sql file to start with 50.